Improve wiring layout for right-hand side.
This is an improvement based on experience wiring the left-hand side.
Wires the rows to the nearest pins to keep from having crossing wires.
Wires the columns to the rest as the wires can just traverse down the
width of the board and find a nice place to intersect the column.

Moves the row-gpios out of the common section since they differ
between halves now.  Updates the right-hand col-gpios.