Revert "Swap command and brackets layer. (#29)"
This reverts commit e20ab5f28bd928ea77687ecf6f9d5dd01d64fd34.

This wasn't working well.  I think this is still something that
I want to do in the future but I need to work on the timing of
the modifiers.

For now, go back to something that was working as I have new
things to implement and I want to try out the builder